Extract ShellCommandRunner from ShellCommand.
